+++ core/vm/interpreter.go
@@ -185,30 +185,6 @@ func (in *WASMInterpreter) CanRun(code []byte) bool {
 	return true
 }
 
-// input = RLP([funcName][params])
-func ParseInputForFuncName(rlpData []byte) (funcName string, err error) {
-	if rlpData == nil || len(rlpData) == 0 {
-		return "", errEmptyInput
-	}
-
-	ptr := new(interface{})
-	rlp.Decode(bytes.NewReader(rlpData), &ptr)
-	rlpList := reflect.ValueOf(ptr).Elem().Interface()
-	if _, ok := rlpList.([]interface{}); !ok {
-		return "", errInvalidRlpFormat
-	}
-
-	iRlpList := rlpList.([]interface{})
-	if len(iRlpList) < 1 {
-		return "", errInsufficientParams
-	}
-
-	if v, ok := iRlpList[0].([]byte); ok {
-		funcName = string(v)
-	}
-	return
-}
-
 // input = RLP([params])
 // returnType must void
 func ParseInitFunctionByABI(vm *exec.VirtualMachine, input []byte, abi []byte) (params []int64, returnType string, err error) {
