+++ core/vm/interpreter_test.go
@@ -135,40 +135,6 @@ func TestWASMInterpreter_Run_Error(t *testing.T) {
 	assert.Equal(t, errInvalidRlpFormat, err)
 }
 
-func TestParseInputForFuncName(t *testing.T) {
-	funcName, err := ParseInputForFuncName(nil)
-	assert.Equal(t, "", funcName)
-	assert.Equal(t, errEmptyInput, err)
-
-	funcName, err = ParseInputForFuncName([]byte{})
-	assert.Equal(t, "", funcName)
-	assert.Equal(t, errEmptyInput, err)
-
-	funcName, err = ParseInputForFuncName([]byte{1, 2, 3})
-	assert.Equal(t, "", funcName)
-	assert.Equal(t, errInvalidRlpFormat, err)
-
-	input, err := rlp.EncodeToBytes([]interface{}{})
-	funcName, err = ParseInputForFuncName(input)
-	assert.Equal(t, "", funcName)
-	assert.Equal(t, errInsufficientParams, err)
-
-	input, err = rlp.EncodeToBytes([]interface{}{""})
-	funcName, err = ParseInputForFuncName(input)
-	assert.Equal(t, "", funcName)
-	assert.NoError(t, err)
-
-	input, err = rlp.EncodeToBytes([]interface{}{"funcName"})
-	funcName, err = ParseInputForFuncName(input)
-	assert.Equal(t, "funcName", funcName)
-	assert.NoError(t, err)
-
-	input, err = rlp.EncodeToBytes([]interface{}{"funcName", ""})
-	funcName, err = ParseInputForFuncName(input)
-	assert.Equal(t, "funcName", funcName)
-	assert.NoError(t, err)
-}
-
 func TestParseInitFunctionByABI(t *testing.T) {
 	lifeVm := &exec.VirtualMachine{}
 	num := utils.Uint64ToBytes(100)
