+++ core/vm/resolver/instructions.go
@@ -38,7 +38,7 @@ func (r *Resolver) envSetState(vm *exec.VirtualMachine) int64 {
 	copy(copyKey, vm.Memory.Memory[key:key+keyLen])
 	copy(copyValue, vm.Memory.Memory[value:value+valueLen])
 
-	log.Info("Get Params From Memory ", "copyKey", string(copyKey), "copyValue", copyValue)
+	log.Info("Get Params From Memory ", "address", r.Service.Address(), "copyKey", string(copyKey), "copyValue", copyValue)
 	r.Service.StateDBService.SetState(r.Service.Address(), copyKey, copyValue)
 	return 0
 }
